# superheroes/superhero-03-01/backend/routers/chat.py
from fastapi import APIRouter, HTTPException
from pydantic import BaseModel
import requests
import os
from bson import ObjectId
from backend.back_server import db_helper
from ..JSONValidation.validators import new_chat_validator, ValidationError
import datetime
from fastapi.responses import JSONResponse


router = APIRouter()

class NewMessage(BaseModel):
    authorName: str
    body: str
    timestamp: str
    isDeleted: bool

class Message(BaseModel):
    currentConversation: str
    newMessage: NewMessage

# def createGeminiContext(messages):
#     return " ".join([message['body'] for message in messages])

def createGeminiContext(chat, new_message):
    if 'pinnedMessages' in chat[0] and len(chat[0]['pinnedMessages']) > 0:
        print(f"ERROR: Pinned messages exist but logic to handle is not implemented yet.")
        return "ERROR: Pinned messages exist but logic to handle is not implemented yet."
    else:  # No pinned messages - Context is done from messages only.
        complete_messages = chat[0]['messages'] + [new_message]
        context = (
            "I am working on requirements engineering. I would like original features suggestions based on existing project requirements.\n"
            "Next is the current state of my suggestions in a conversation format. I would like you to answer the latest message while taking into account the previous conversation.\n"
            "If you are suggesting new features, your answer should outline them so that it is clear what they are.\n"
            "If the latest message is a question, your answer should be a short and clear response only to that question but taking into account the conversation if necessary.\n"
        )
        for message in complete_messages:
            context += f"\n{message['authorName']}: {message['body']}"
        return context


@router.post("/")
async def send_message(message: Message):
    # Validate the request body
    try:
        new_chat_validator(message.model_dump())
    except ValidationError as e:
        # Log to the terminal
        print(f"ERROR:  POST /chat/ Bad request: {e.message}")
        raise HTTPException(status_code=400, detail=f"Bad request: {e.message}")
    
    try:
        LLM_API_KEY = os.getenv("LLM_API_KEY")
        if not LLM_API_KEY:
            raise ValueError("LLM_API_KEY not found in environment variables")

        success, chat = db_helper.getChat(message.currentConversation)
        
        if success:
            # complete_messages = chat[0]['messages'] + [message.newMessage.model_dump()]
            context = createGeminiContext(chat, message.newMessage.model_dump())
        else:
            context = message.newMessage.body

        # Send the context to the gemini
        headers = {
            "Content-Type": "application/json",
        }
        gemini_payload = {
            "contents": [
                {
                    "parts": [
                        {
                            "text": context
                        }
                    ]
                }
            ]
        }

        # gemini_api_url = f"https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash-latest:generateContent?key={LLM_API_KEY}"
        gemini_api_url = f"https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash-8b:generateContent?key={LLM_API_KEY}"

        gemini_response = requests.post(gemini_api_url, json=gemini_payload, headers=headers)
        if gemini_response.status_code != 200:
            raise HTTPException(status_code=gemini_response.status_code, detail="Currently unable to generate response. Error with Gemini API.")

        response_message = gemini_response.json().get("candidates", [{}])[0].get("content", {}).get("parts", [{}])[0].get("text", "No response from LLM")

        # Construct the message object
        new_message = {
            "authorName": "Gemini",
            "body": response_message,
            "timestamp": datetime.datetime.now().isoformat(),
            "isDeleted": False
        }
        
        # If everything is successful, update the database
        if success: # If the chat already exists
            new_messages = [message.newMessage.model_dump() , new_message]
            db_helper.addMessagesToChat(message.currentConversation, new_messages)
            
            return JSONResponse(content=new_message, status_code=200)
        
        else: # If the chat is new
            conversationId = db_helper.createChat(message.newMessage.model_dump())
            db_helper.addMessagesToChat(conversationId, [new_message])

            # Get chat to send in the response
            success, chat = db_helper.getChat(conversationId)
            if success:
                chat[0]['_id'] = str(chat[0]['_id'])
                return JSONResponse(content=chat[0], status_code=201) # Explicit success code to be used for behaviour in the frontend

    except Exception as e:
        # Clean up the database if something goes wrong
        # TODO
        raise HTTPException(status_code=500, detail=f"Internal server error: {str(e)}")

@router.get("/pin/{id}")
async def pin_message_by_id(id: str):
    try:
        exists, chat = db_helper.getChat(id)
        if not exists:
            raise HTTPException(status_code=404, detail="Chat not found")

        pinned_message = chat['pinnedMessages']
        return pinned_message

    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Internal server error: {str(e)}")

class PinMessageRequest(BaseModel):
    message: str

@router.post("/pin/{id}")
async def pin_message_by_id(id: str, body: PinMessageRequest):
    try:
        # Use the message from the request body
        exists, chat = db_helper.addPinnedToChat(id, body.message)
        print(chat)
        if not exists:
            raise HTTPException(status_code=404, detail="Chat not found")

        pinned_message = chat['pinnedMessages']
        return pinned_message

    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Internal server error: {str(e)}")

@router.get("/{id}")
async def get_message_by_id(id: str):
    try:
        exists, chat = db_helper.getChat(id)
        if not exists:
            raise HTTPException(status_code=404, detail="Chat not found")
        return chat

    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Internal server error: {str(e)}")